requirements.txt
Flask>=2.2
requests
python-dotenv
numpy
pandas
scikit-learn
joblib
gunicorn
